Refactor Webpack config to be easier to customise in more advanced projects
