+++ webpack/webpack.config.dev.js
@@ -1,20 +1,10 @@
-const webpack = require('webpack');
-const base = require('./webpack.config.js');
+const base = require('./webpack.config.base');
 
-/**
- * Development Webpack config, with dev-workflow optimisations.
- */
-module.exports = Object.assign({}, base, {
+const config = base('development');
+
+module.exports = Object.assign({}, config, {
+    // See http://webpack.github.io/docs/configuration.html#devtool
     devtool: 'inline-source-map',
-
-    plugins: base.plugins.concat([
-        new webpack.DefinePlugin({
-            'process.env': {
-                NODE_ENV: JSON.stringify('development'),
-            },
-        }),
-    ]),
-
     // Turn off performance hints during development because we don't do any
     // splitting or minification in interest of speed. These warnings become
     // cumbersome.
